Move AddDofInterpolation to separate files
Mainly to remove dependencies from ANN_LIBRARIES if this feature is
unwanted.
