+++ src/mechanics/CMakeLists.txt
@@ -143,6 +143,8 @@ set(MechanicsMeshSources
     mesh/MeshCompanion.cpp
     mesh/MeshGenerator.cpp
     mesh/MeshFem.cpp
+    mesh/MeshFemDofConvert.cpp
+    mesh/UnitMeshFem.cpp
     )
 
 set(MechanicsNodesSources
