+++ test/mechanics/mesh/CMakeLists.txt
@@ -9,10 +9,10 @@ endif()
 add_test(almost::mechanics::mesh::MeshCompanion ${CMAKE_CURRENT_BINARY_DIR}/MeshCompanion)
 
 add_unit_test(MeshFem
+    mechanics/mesh/MeshFemDofConvert
     mechanics/elements/ElementShapeFunctions)
 target_link_libraries(MeshFem ${ANN_LIBRARIES})
 
 add_unit_test(UnitMeshFem
     mechanics/mesh/MeshFem
     mechanics/elements/ElementShapeFunctions)
-target_link_libraries(UnitMeshFem ${ANN_LIBRARIES})
